move email DBFunctions to functions namespace
